pyproject.toml
[tool.poetry]
name = "ena"
version = "0.1.0"
description = ""
authors = ["arkhon7 <reviuy9@gmail.com>"]

[tool.black]
line-length = 120

[tool.poetry.dependencies]
python = "3.9.6"
hikari-lightbulb = "^2.2.0"
python-dotenv = "^0.19.2"
simpleeval = "^0.9.12"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.39"}
hikari = "^2.0.0-alpha.109"
hikari-miru = "^1.1.1"
psycopg2 = "^2.9.3"
asyncpg = "^0.26.0"
aiocache = "^0.11.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.1.0", allow-prereleases = true}
pylint = "^2.12.2"
mypy = "^0.931"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
